requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==20.2.0
autopep8==1.5.4
coverage==5.3
Django==3.1.2
django-extensions==3.0.9
django-nested-admin==3.3.2
django-reverse-admin==2.9.0
django-shell-plus==1.1.7
djangorestframework==3.12.1
Faker==0.9.1
importlib-metadata==2.0.0
iniconfig==1.0.1
isort==5.5.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mixer==6.1.3
mock==4.0.2
packaging==20.4
pip==9.0.1
pkg-resources==0.0.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.1
pytest-cov==2.10.1
pytest-django==3.10.0
python-dateutil==2.8.1
python-monkey-business==1.0.0
pytz==2020.1
robotframework==3.2.2
robotframework-pythonlibcore==2.1.0
robotframework-seleniumlibrary==4.5.0
selenium==3.141.0
setuptools==39.0.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.2
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.11
wrapt==1.12.1
zipp==3.3.0
